+++ debian12/scripts/setup_avahi.sh
@@ -1,4 +1,4 @@
-#!/bin/bash
+#!/usr/bin/env bash
 
 if [ -e /usr/sbin/avahi-daemon ] && [ -e /var/run/dbus ]
 then