+++ debian12/scripts/setup_iob_db.sh
@@ -1,4 +1,4 @@
-#!/bin/bash
+#!/usr/bin/env bash
 
 # reading env
 debug=$DEBUG